requirements.txt
amqp==5.0.6
APScheduler==3.7.0
asgiref==3.3.1
atomicwrites==1.4.0
attrs==21.1.0
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.1.7
django-allauth==0.44.0
django-apscheduler==0.5.2
django-bootstrap-themes==3.3.6
django-celery-results==2.0.1
django-ckeditor==6.0.0
django-js-asset==1.2.2
django-redis==4.12.1
flake8==3.9.1
greenlet==1.0.0
idna==2.10
iniconfig==1.1.1
kombu==5.0.2
mccabe==0.6.1
numpy==1.19.2
oauthlib==3.1.0
packaging==20.9
Pillow==7.2.0
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2==2.8.6
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyJWT==2.0.1
pyparsing==2.4.7
PyQt5==5.15.1
PyQt5-sip==12.8.1
pytest==6.2.4
pytest-django==4.2.0
python3-openid==3.2.0
pytz==2020.4
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sorl-thumbnail==12.7.0
SQLAlchemy==1.4.9
sqlparse==0.4.1
toml==0.10.2
tzlocal==2.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
wxPython==4.1.0
